🐛 fix: 기존 모달과 수정한 모달 파일명 변경
- 기존 모달은 LegacyModal.tsx로, 수정한 모달은 index.tsx로 파일명 변경
